Don't add `createPage` until needed. (#3354)
Based on the tutorial the `createPage` method is only needed when you actually create the pages. So this tutorial change removes it from the previous step which just `console.logs` the page the nodes and then highlights that you need to add it in the next step where `createPage` is then actually used.

Tripped me as I was following the tutorial since my editor complained that it was an unused import.