Explained what cd means for starter installation (#4737)
* Explained what cd means for starter installation

Is "change" the right word in this phrasing?

"Similar to part one, open a new terminal window and run the following commands to install and then change to your new tutorial-part-two directory:"

* made name of directory clear

* Add back ticks to directory, grammar edit

* Add back ticks to directory, grammar

Broke up the sentence and opted for “just like in” for clarity. “Similar to” makes me have to think “what is similar?” which makes me doubt myself.

* added back ticks

Thanks for this idea!