feat(examples): Add "Edit on Codesandbox" button (#11409)
As per discussion with @m-allanson we'll add two "edit on codesandbox" buttons in the READMEs of the examples. These examples/sources get linked on several doc pages so it's not totally necessary to change the docs about these, too.
We can track with the views/forks how many people use this offer.