docs(gatsby-plugin-typography): New tutorial link (#13300)
## Description

Typography.js is now in tutorial part 3, not part 2.

(The previous link was to an anchor within part 2, but this link is to the top of part 3. I chose not to link to [Using Plugins](https://www.gatsbyjs.org/tutorial/part-three/#using-plugins) because it's not explicitly a section about Typography.js, because it's not far into the document, and because I think the preamble provides useful context. But I could have gone either way.)