chore(gatsby): migrate status reducer to TypeScript (#23474)
* chore(gatsby): migrate status reducer to TypeScript

This commit migrates the status reducer to TypeScript by adding the
relevant types for each actions.

In order to make sense of the condition protecting the
`SET_PLUGIN_STATUS` action, I had to transform the `&&` operator to
`||`. Without this change, the typecheck would fail. But in the end,
it should not change the behaviour of this early-termination.

* chore(gatsby): improve readability of ActionsUnion

With this commit, the order of the actions interfaces taking part
of the ActionsUnion type is updated to be alphabetical.

The goal of this move is to improve the readability and the
maintainability of what starts to be a huge type union.