feat(gatsby-recipes): Implement initial custom react renderer for recipes (#24065)
* feat(gatsby-recipes): Implement initial custom react renderer for recipes

This implements the first pass of a custom React reconciler and renderer for
Recipes. The idea is that we render a recipe directly and construct a plan.
The rendered components related to the resource suspend and asynchronously
call the resource plan which is then cached. When rendering, all resource
calls are added to an array and we `await` their completion before calling
render again which returns the desired data.

We also tested it with a conditional component and it worked!

    function ConditionalFile(props) {
      if (props.path === 'old') {
        return <File {...props} path='new'
      }

      return <File {...props} />
    }

Related: https://github.com/gatsbyjs/gatsby/issues/22991#issuecomment-625985760

---

Co-authored-by: Kyle Mathews <mathews.kyle@gmail.com>

* feat(gatsby-recipes): Implement a basic jsx renderer

A recipes comes in as MDX which we then transpile to JSX, however
we need an actual React element which we can pass to the renderer.
We achieve this be using the function constructor after a babel
transform for React.createElement.

Additionally, we now transform MDX JSX into the raw JSX string
that the renderer expects.

* feat(gatsby-recipes): Wire up renderer to recipes core

Updates the recipes state machine and client to use a simplified
data structure. Also uses the new custom react renderer to create
the data.

Co-authored-by: Kyle Mathews <mathews.kyle@gmail.com>

* fix(gatsby-recipes): Update transformer test for new data shape

* fix(gatsby-recipes): Update tests, handle nested JSX

Co-authored-by: Kyle Mathews <mathews.kyle@gmail.com>

* fix(gatsby-recipes): Add parse validation for JSX blocks

* fix(gatsby-recipes): Ensure babel uses built ins, make Config a noop

* fix(gatsby-recipes): Add back validation for no resource in first step

* fix(gatsby-recipes): Handle schema validation errors at render time

* chore(gatsby-recipes): Remove unused deps

* fix(gatsby-recipes): Update snapshots

Co-authored-by: Kyle Mathews <mathews.kyle@gmail.com>