chore(gatsby): migrate run-sift to TS (#25055)
* Rename file to .ts and drop the sift name

This change is broken because references are not updated. Separate commit for your reviewing pleasure. (The PR will squash it and github will still screw up the rename but at least review is good.)

* Convert run-fast-filters (run-sift) to TS

* Applied suggestions

* Remove more unknowns

* wip

* Fix regression

While trying to be type correct, coercing the value to string has the unintended side effect of allowing regexes (or any other type) as $regex arg. So instead now throwing an explciit error for that case.

Also reverted the type where it returns an Array. Apparently there are cases where it is not an array.