feat(gatsby-source-graphql): Add transformSchema option (#25048)
* [gatsby-source-graphql] add transfomSchema option

Should resolve #23552. Adds an optional custom transfomSchema option to gatsby-source-graphql that passes the default transforms, schema and link. I has many potential usages, but the consumer should be able to use the same default transforms or tweak or replace them completely. The original use case is to add an argument to the generated "fieldName" field that can then be read by a custom apollo-link to customize the underlying http call with a particular header. Tagging @vladar. This is a proposed implementation that should be safe when not used. I'm not sure if other parameters might be passed along the transformSchema arguments, maybe plugin options?

* Fix implementation, renamed option

* Added options parameter and readme example

* Removed example + default implementation + references

* Renamed transformSchema in customWrapSchemaFn

* Renamed back option into "transformSchema"

Co-authored-by: Stefano De Pace <depace@ARIADNEAD.DOM>
Co-authored-by: gatsbybot <mathews.kyle+gatsbybot@gmail.com>